+++ assets/utilities/element.util.mjs
@@ -1,0 +1,8 @@
+export function setClass(className, e){
+    e.classList.add(className);
+}
+export function resetClass(className, e){
+    e.classList.remove(className);
+}
+
+
